refactor: Set NewHomePage as default, remove old Index.tsx
- Moved mock data (programs, advertisements) and related interfaces from the old Index.tsx to a new central file `src/data/mockData.ts`.
- Updated all components and contexts that used this mock data to import from the new location.
- Updated routing in `App.tsx` to make `NewHomePage` the component for the root path (`/`).
- Updated navigation links in `MainLayout.tsx` to point to the new homepage at `/`.
- Deleted the old `src/pages/Index.tsx` file as it's no longer needed.
